+++ src/main/java/com/untouchable/everytime/Book/Entity/BookEntity.java
@@ -1,0 +1,25 @@
+package com.untouchable.everytime.Book.Entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.sql.Date;
+
+@Entity
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class BookEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long book_PK;
+    String title;
+    String author;
+    String publisher;
+    Date publicationDate;
+    Long bookPrice;
+}
