+++ src/main/java/com/untouchable/everytime/Book/Enum/BookStatus.java
@@ -1,0 +1,5 @@
+package com.untouchable.everytime.Book.Enum;
+
+public enum BookStatus {
+    NONE, PENCIL, BALL_POINT
+}
