--- src/main/java/com/untouchable/everytime/DTO/MailBlockDTO.java
@@ -1,18 +1,0 @@
-package com.untouchable.everytime.DTO;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class MailBlockDTO {
-
-    Long mailBlock_PK;
-
-    Long user_PK;
-
-}
