--- src/main/java/com/untouchable/everytime/DTO/MailDTO.java
@@ -1,23 +1,0 @@
-package com.untouchable.everytime.DTO;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.sql.Date;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class MailDTO {
-
-    Long mail_PK;
-
-    String reciever;
-    String sender;
-    String comment;
-    Date createdAT;
-
-}
