--- src/main/java/com/untouchable/everytime/DTO/UserPointHistoryDTO.java
@@ -1,21 +1,0 @@
-package com.untouchable.everytime.DTO;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-
-public class UserPointHistoryDTO {
-
-    Long userPoint_PK;
-
-    Long user_PK;
-
-    Long point;
-
-}
