--- src/main/java/com/untouchable/everytime/Entity/BookEntity.java
@@ -1,25 +1,0 @@
-package com.untouchable.everytime.Entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.sql.Date;
-
-@Entity
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class BookEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long book_PK;
-    String title;
-    String author;
-    String publisher;
-    Date publicationDate;
-    Long bookPrice;
-}
