--- src/main/java/com/untouchable/everytime/Entity/ScheduleEntity.java
@@ -1,34 +1,0 @@
-package com.untouchable.everytime.Entity;
-
-import com.untouchable.everytime.Lecture.Entity.LectureEntity;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.experimental.SuperBuilder;
-
-@Entity
-@SuperBuilder
-@NoArgsConstructor
-@AllArgsConstructor
-@Data
-public class ScheduleEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long schedulePK;
-
-    @ManyToOne
-    LectureEntity lecture;
-
-    String location;
-
-    // 일요일 0 ~ 토요일 6
-    int day;
-
-    // 0시 ~ 24시 , start < end
-    int start;
-    int end;
-
-
-
-}
