--- src/main/java/com/untouchable/everytime/Entity/UserAppliedLecture.java
@@ -1,4 +1,0 @@
-package com.untouchable.everytime.Entity;
-
-public class UserAppliedLecture {
-}
