--- src/main/java/com/untouchable/everytime/Enum/BookStatus.java
@@ -1,5 +1,0 @@
-package com.untouchable.everytime.Enum;
-
-public enum BookStatus {
-    NONE, PENCIL, BALL_POINT
-}
