+++ src/main/java/com/untouchable/everytime/Schedule/Entity/UserAppliedLecture.java
@@ -1,0 +1,4 @@
+package com.untouchable.everytime.Schedule.Entity;
+
+public class UserAppliedLecture {
+}
